Trying to sort Shape before offset
