+++ Shape.py
@@ -8,6 +8,7 @@ import Line as l
 import Point as p
 from LineGroup import LineGroup as LG
 from parameters import constants as c
+import parameters as pr
 
 class Shape(LG):    
     def __init__(self, shape):
@@ -51,12 +52,22 @@ class Shape(LG):
     def offset(self, distance, desiredSide):
         trimJoin = self.trimJoin_coro()
         next(trimJoin)
-        for line in self:
-            try1, try2 = line.getOffsetLines(distance)
-            if self.isInside(try1.getMidPoint()) is desiredSide:
-                trimJoin.send(try1)
+        nearestCoro = pr.nearestLine_Coro(self)
+        next(nearestCoro)
+        testPoint = p.Point(0,0)
+        while True:
+            try:
+                line, _, _ = nearestCoro.send((True, testPoint))
+            except StopIteration:
+                break
             else:
-                trimJoin.send(try2)
+                testPoint = line.end
+                try1, try2 = line.getOffsetLines(distance)
+                if self.isInside(try1.getMidPoint()) is desiredSide:
+                    trimJoin.send(try1)
+                else:
+                    trimJoin.send(try2)
+        nearestCoro.close()
         return Shape(trimJoin.send(None))
        
     def trimJoin_coro(self):
